+++ PokemonEncyclopedia/PokemonEncyclopediaTests/FeedCache/CacheFeedUseCaseTests.swift
@@ -133,21 +133,4 @@ class CacheFeedUseCaseTests: XCTestCase {
 		XCTAssertEqual(receivedError as NSError?, expectedError, file: file, line: line)
 	}
 
-	private func uniqueImage() -> FeedImage {
-		return FeedImage.init(id: UUID(), name: "ane", url: anyURL())
-	}
-	
-	private func uniqueImageFeed() -> (models: [FeedImage], local: [LocalFeedImage]) {
-		let models = [uniqueImage(), uniqueImage()]
-		let local = models.map {LocalFeedImage(id: $0.id, name: $0.name, url: $0.url)}
-		return (models, local)
-	}
-	
-	private func anyURL() -> URL{
-		return URL.init(string: "http://any-url.com")!
-	}
-
-	private func anyNSError() -> NSError{
-		return  NSError(domain: "any error", code: 0, userInfo: nil)
-	}
 }
