+++ PokemonEncyclopedia/PokemonEncyclopediaTests/FeedCache/LoadFeedFromCacheUseCaseTests.swift
@@ -179,31 +179,4 @@ class LoadFeedFromcacheUseCaseTest: XCTestCase {
 		wait(for: [exp], timeout: 1.0)
 	}
 
-	private func uniqueImage() -> FeedImage {
-		return FeedImage.init(id: UUID(), name: "ane", url: anyURL())
-	}
-	
-	private func uniqueImageFeed() -> (models: [FeedImage], local: [LocalFeedImage]) {
-		let models = [uniqueImage(), uniqueImage()]
-		let local = models.map {LocalFeedImage(id: $0.id, name: $0.name, url: $0.url)}
-		return (models, local)
-	}
-	
-	private func anyURL() -> URL{
-		return URL.init(string: "http://any-url.com")!
-	}
-	
-	private func anyNSError() -> NSError{
-		return  NSError(domain: "any error", code: 0, userInfo: nil)
-	}
-}
-
-
-private extension Date {
-	func adding(days: Int) -> Date {
-		return Calendar(identifier: .gregorian).date(byAdding: .day, value: days, to: self)!
-	}
-	func adding(seconds: TimeInterval) -> Date {
-		return self + seconds
-	}
 }
